Move theme getting from React app to the index file
